extract voice chatbot
